Chorgan v0.2.5 — fix sineSample overflow causing crash at SINE zone
sineSample() computed: (s2 * frac + s1 * (65536 - frac)) >> 16
with s1/s2 up to ±32000 and frac up to 65536. Each term can reach ~2.1B
and their sum overflows int32_t (max 2.147B) — signed UB on Cortex-M0+.
This caused a CPU fault specifically in the SINE blend zone (Main knob
fully CCW/CW), corrupting processor state that survived warm resets.

Fix: promote both multiply operands to int64_t before the addition.

+++ Demonstrations+HelloWorlds/PicoSDK/ComputerCard/examples/chorgan/main.cpp
@@ -399,7 +399,7 @@ inline int32_t ChorganCard::sineSample(uint32_t ph) {
     uint32_t frac  = (ph & 0x7FFFFFu) >> 7;
     int32_t s1 = SINE_TABLE[index & 0x1FFu];
     int32_t s2 = SINE_TABLE[(index + 1u) & 0x1FFu];
-    return (s2 * (int32_t)frac + s1 * (int32_t)(65536u - frac)) >> 16;
+    return (int32_t)(((int64_t)s2 * frac + (int64_t)s1 * (65536u - frac)) >> 16);
 }
 
 inline int32_t ChorganCard::triSample(uint32_t ph) {
